\name{CladeBrownFrogsCoordsLocs}
     \docType{data}
     \alias{CladeBrownFrogsCoordsLocs}
     \title{Coordinates and haplotypes for the Brown Frog dataset}
     \description{
Dataset of all known taxa of the "western" and "eastern" lineage of the Near Eastern brown frogs (\emph{Rana macrocnemis macrocnemis, R. m. camerani, R. holtzi, R. m. pseudodalmatina,  R. m. tavasensis}, Amphibia, Ranidae). Contains the coordinates of each sampling location along with the haplotypes found in each location. It is called using \code{data(CladeBrownFrogsCoordsLocs)}.
     }
     \usage{CladeBrownFrogsCoordsLocs}
     \format{In total 40 sampling locations throughout the Near Eastern (Turkey, Georgia, Daghestan, Northern Iran) with 21 (19 + 2) haplotypes.
 }
    
     \references{
M. Veith, J.F. Schmidtler, J. Kosuch, I. Baran and A. Seitz (2003) Palaeoclimatic changes explain Anatolian mountain frog evolution: a test for alternating vicariance and dispersal events. Molecular Ecology. 

M. Veith, J. Kosuch and M. Vences (2003) Climatic oscillations triggered post-Messinian speciation of Western Palearctic brown frogs (Amphibia, Ranidae). Molecular Phylogenetics and Evolution. 

     }
     \examples{
#to use the example dataset:
data(CladeBrownFrogsCoordsLocs)
coordsLocs <- CladeBrownFrogsCoordsLocs
}

     \keyword{Brown Frog}
