\name{CladeBrownFrogsRawSeqs}
     \docType{data}
     \alias{CladeBrownFrogsRawSeqs}
     \title{Sequences for the Brown Frog dataset}
     \description{Dataset of all known taxa of the "western" and "eastern" lineage of the Near Eastern brown frogs (\emph{Rana macrocnemis macrocnemis, R. m. camerani, R. holtzi, R. m. pseudodalmatina,  R. m. tavasensis, Amphibia, Ranidae, R. m. tavasensis}, Amphibia, Ranidae) for the exploration of deeper phylogeographic patterns.
       Contains the 16S ribosomal mtDNA sequences (n=19) listed in the Appendix II of Veith et al (2003). It is called using \code{data(CladeBrownFrogsRawSeqs)}.
     }
     \usage{CladeBrownFrogsRawSeqs}
     \format{In total 40 sampling locations throughout the Near Eastern (Turkey, Georgia, Daghestan, Northern Iran) with 21 (19 + 2) haplotypes of length 536.}
  
     \references{
M. Veith, J.F. Schmidtler, J. Kosuch, I. Baran and A. Seitz (2003) Palaeoclimatic changes explain Anatolian mountain frog evolution: a test for alternating vicariance and dispersal events. Molecular Ecology. 

M. Veith, J. Kosuch and M. Vences (2003) Climatic oscillations triggered post-Messinian speciation of Western Palearctic brown frogs (Amphibia, Ranidae). Molecular Phylogenetics and Evolution. 

     }
     
       \examples{
#to use the example dataset:
data(CladeBrownFrogsRawSeqs)
rawSeqs <- CladeBrownFrogsRawSeqs
}
     \keyword{Brown Frog}
