\name{MacrocnemisRawSeqs}
\encoding{ISO-8859-1}
     \docType{data}
     \alias{MacrocnemisRawSeqs}
     \title{Sequences for the R.macrocnemis Brown Frog dataset}
     \description{\emph{Rana macrocnemis} (Amphibia, Ranidae) dataset, the Near Eastern brown frog, from a transition zone of landscapes in Georgia, Caucasus.
       Contains the mtDNA sequences (cytochrome b)  of the Brown Frog dataset used in the BPEC examples. It is called using \code{data(MacrocnemisRawSeqs)}.
     }
     \usage{MacrocnemisRawSeqs}
     \format{In total 40 mtDNA sequences of length 504 nucleotides each.}
  
     \references{
D. Tarkhnishvili, A. Hille  and W. B?hme (2001)
Humid forest refugia, speciation and secondary introgression between evolutionary lineages: differentiation in a Near Eastern brown frog, Rana macrocnemis. Biological Journal of the Linnean Society.

      A. Hille, I. Manolopoulou and D. Tarkhnisvili (2015)
Detecting landscape-dependent selection with Bayesian phylogeographic and ecological clustering: 
comparison of recombinant microsatellite and mtDNA haplotype data,
(unpublished).
     }
     \examples{
#to use the example dataset:
data(MacrocnemisRawSeqs)
rawSeqs <- MacrocnemisRawSeqs
}
     \keyword{Brown Frog}
