\name{TransalpinaCoordsLocs}
     \docType{data}
     \alias{TransalpinaCoordsLocs}
     \title{Coordinates and haplotypes for the burnet moth dataset.}
     \description{The dataset of three closely related forms of a European burnet moth with a complicated micro-vicariant distribution pattern, the \emph{Zygaena transalpina} complex (Lepidoptera, Zygaenidae). Contains the coordinates of each sampling location along with the haplotypes found in each location. It is called using data(TransalpinaCoordsLocs).}
     \usage{TransalpinaCoordsLocs}
     \format{In total 55 sampling locations with rich population samples per location revealing plenty of distinct haplotypes.}
    \author{compiled by Axel Hille using supplementary material of von Reumont et al (2012). }
     \references{
B.M. von Reumont, J.-F. Struwe, J. Schwarzer and B. Misof (2012)  Phylogeography of the burnet moth Zygaena transalpina-complex: molecular and morphometrical differentiation suggests glacial refugia in Southern France, Western Europe and micro refugia within the Alps. Journal of Zoological Systematics and Evolutionary Research.      
     }
     \examples{
#to use the example dataset:
data(TransalpinaCoordsLocs)
coordsLocs <- TransalpinaCoordsLocs
}
     \keyword{Burnet Moth}
